requirements.txt
Django>=4.2
djangorestframework>=3.14
black>=23.9.1
isort>=5.12.0
flake8>=6.0.0
